requirements.txt
selenium==4.30.0
pillow==11.0.0
pytesseract==0.3.10
numpy==2.0.2
opencv-python-headless==4.11.0.86
matplotlib==3.9.2
cairosvg==2.8.0
beautifulsoup4==4.13.3
lxml==5.4.0
webdriver-manager==4.0.2 
requests==2.32.0
boto3
